requirements.txt
PyQt5>=5.15.0
torch>=2.4.0
torchaudio>=2.4.0
torchvision>=0.19.0
transformers>=4.57.0
sentencepiece>=0.1.99
coqui-tts>=0.27.0
soundfile>=0.12.0
pydub>=0.25.0
numpy>=1.26.0
librosa>=0.11.0
scipy>=1.13.0